Reorganize files structure
