+++ packages/teleterm/src/services/pty/index.ts
@@ -1,0 +1,2 @@
+export * from './ptyService';
+export * from './types';
