+++ packages/teleterm/src/services/pty/ptyHost/resolveShellEnv.ts
