--- packages/teleterm/src/sharedProcess/ptyHost/client/fixtures/mocks.ts
@@ -1,29 +1,0 @@
-import { PtyProcess, PtyServiceClient } from '../../types';
-
-export class MockPtyProcess implements PtyProcess {
-  start() {}
-
-  write() {}
-
-  resize() {}
-
-  dispose() {}
-
-  onData() {}
-
-  onExit() {}
-
-  onOpen() {}
-
-  getPid() {
-    return 0;
-  }
-
-  getCwd = async () => '';
-}
-
-export class MockPtyServiceClient implements PtyServiceClient {
-  createPtyProcess(): Promise<PtyProcess> {
-    return Promise.resolve(new MockPtyProcess());
-  }
-}
