--- packages/teleterm/src/sharedProcess/ptyHost/client/index.ts
@@ -1,1 +1,0 @@
-export * from './ptyHostClient';
