+++ packages/teleterm/src/sharedProcess/ptyHost/index.ts
@@ -1,1 +1,3 @@
 export * from './types';
+export * from './v1/ptyHostService_grpc_pb';
+export * from './v1/ptyHostService_pb';
