--- packages/teleterm/src/sharedProcess/ptyHost/server/index.ts
@@ -1,1 +1,0 @@
-export * from './ptyHostService';
