+++ packages/teleterm/src/types.ts
@@ -1,5 +1,5 @@
 import { TshClient } from 'teleterm/services/tshd/types';
-import { PtyServiceClient } from 'teleterm/sharedProcess/ptyHost';
+import { PtyServiceClient } from 'teleterm/services/pty';
 import { RuntimeSettings, MainProcessClient } from 'teleterm/mainProcess/types';
 import { Logger, LoggerService } from './services/logger/types';
 import { FileStorage } from 'teleterm/services/fileStorage';
