+++ packages/teleterm/src/ui/DocumentTerminal/Terminal/Terminal.tsx
@@ -18,7 +18,7 @@ import { debounce } from 'lodash';
 import React, { useEffect, useRef } from 'react';
 import styled, { useTheme } from 'styled-components';
 import { Box, Flex } from 'design';
-import { PtyProcess } from 'teleterm/sharedProcess/ptyHost';
+import { PtyProcessType } from 'teleterm/sharedProcess/ptyHost';
 import XTermCtrl from './ctrl';
 
 export default function Terminal(props: Props) {
@@ -74,7 +74,7 @@ export default function Terminal(props: Props) {
 }
 
 type Props = {
-  ptyProcess: PtyProcess;
+  ptyProcess: PtyProcessType;
   visible: boolean;
   onEnterKey?(): void;
 };
